Рефакторинг. Переместил кучу кода в отдельный класс. Вместо int теперь стоит long в нужных местах

+++ src/main/java/cryptotoolsomsu/Utils.java
@@ -7,22 +7,22 @@ import java.util.ArrayList;
 public class Utils {
 
     // Алгоритм найдет все простые множители n. Необходим для вычисления функции Эйлера по формуле.
-    public static ArrayList<Integer> primeFactor(int n) {
+    public static ArrayList<Long> primeFactor(long n) {
 
         // Создадим массив для результатов.
-        ArrayList<Integer> result = new ArrayList<>();
+        ArrayList<Long> result = new ArrayList<>();
 
         //Этот цикл разделит n на два до тех пор пока не сможем.
         while (n % 2 == 0) {
-            if (!(result.contains(2))) {
-                result.add(2);
+            if (!(result.contains((long) 2))) {
+                result.add((long) 2);
             }
             n = n / 2;
         }
 
         // Этот цикл удостоверяется что i не будет скакать до самого n
         // (тот факт что он не превышает корня из n доказуем)
-        for (int i = 3; i <= Math.sqrt(n); i += 2) {
+        for (long i = 3; i <= Math.sqrt(n); i += 2) {
 
             // Делим на i пока не сможем
             while (n % i == 0) {
@@ -42,18 +42,18 @@ public class Utils {
     }
 
     // Алгоритм для нахождения функции Эйлера
-    public static int eulerFunction(int n) {
+    public static long eulerFunction(long n) {
 
         // Берем все простые множители n
-        ArrayList<Integer> arrayOfPrimitives = primeFactor(n);
+        ArrayList<Long> arrayOfPrimitives = primeFactor(n);
         float eulerFunction = n;
 
         // Пробегаемся по формуле phi(n) = n * (1 - 1/p_1) * (1 - 1/p_2) * ..., где p_i - простой множитель числа n
-        for (Integer arrayOfPrimitive : arrayOfPrimitives) {
+        for (Long arrayOfPrimitive : arrayOfPrimitives) {
             eulerFunction *= (1.0 - (1.0 / (float)arrayOfPrimitive));
         }
 
-        return (int)eulerFunction;
+        return (long) eulerFunction;
     }
 
 }
